feat(gazelle): Gazelle plugin generates py_proto_library (#3057)
Fixes https://github.com/bazel-contrib/rules_python/issues/2994.

Please go over this with a fine-toothed comb! This is my first
contribution to `rules_python` / the gazelle plugin, and while I've
worked in Gazelle before, I'm pretty unfamiliar with the Python plugin's
architecture.

This adds support in the Gazelle plugin for generating
`py_proto_library` rules automatically, if there are any `proto_library`
rules detected in a given package. We do this via a new Gazelle
directive, `python_generate_proto`, which defaults to `true`, and
controls whether these rules are generated.

See the tests in `testdata/directive_python_generate_proto` for
examples.

By default, we source the `py_proto_library` rule from the `@protobuf`
repository. I think this the intended long-term home of the rule? Users
are expected to use `gazelle:map_kind` to change this if need be.

I haven't done anything here to support resolution of imports of
`py_proto_library`. I think this is worth landing first, to save folks
from having to maintain these by hand. But this should lay the
foundation for resolving that in
https://github.com/bazel-contrib/rules_python/issues/1703.

---------

Co-authored-by: Douglas Thor <dougthor42@users.noreply.github.com>