feat(gazelle): Add `include_pytest_conftest` annotation (#3080)
Fixes #3076.

Add a new gazelle annotation `include_pytest_conftest`. When unset
or true, the gazelle behavior is unchanged. When false, gazelle will
*not* inject the `:conftest` dependency to py_test targets.

One of the refactorings that is done to support this is to pass around
an `annotations` struct in `target.targetBuilder`. This will also open
up support for other annotations in the future.

---------

Co-authored-by: Ignas Anikevicius <240938+aignas@users.noreply.github.com>