refactor(gazelle_manifest): print the wrong hash when encountered (#3103)
Ideally developers should always re-run the manifest generator, which
will update the hash for them.

However I've got clients where the CI system is the only place that all
the dependencies can resolve, either because of credentials needed to
access the wheelhouse/PyPI, or because of disk exhaustion from massive
wheels.

Printing the difference allows a red PR to be greened up just by reading
the CI log.