fix(pypi): expose pypi packages only common to all python versions (#3107)
Closes #2921

For a single pypi repo with multiple python versions, `all_requirements`
fail when a pypi package supports Python version A but not version B. In
this case, the pypi package would be included only in requirements lock
file for version A, not in one for version B. However, the failure
occurs since the package is included in `all_requirements` even for
Python version B.
(Minimal reproduction:
https://github.com/dotoleeoak/rules-python-2921-repro)

This happens since `packages` parameter for `hub_repository` targets are
including all packages across all requirement lock files. Instead of
union of packages, intersection of packages for requirement files should
be passed to `packages` and exposed to `all_requirements` macro, so that
those packages are compatible with all Python versions.

---------

Co-authored-by: Ignas Anikevicius <240938+aignas@users.noreply.github.com>